Replace babel-plugin-module-resolver
Now we use babel-plugin-root-import. The reason is that
babel-plugin-module-resolver has a bug that it doesn't pass
the api object to babel.config.js, which prevents us from
configuring the cache.

The package babel-plugin-root-import does what we need
without this problem. However the previous way of specifying
didn't work, which is actually good because having components
starting with `@` is confusing because there are npm modules
that can start with the same letter. Replacing with `~/` makes
better sense because it acts like a home directory.