chore: Move client IDs and the like to .env
No one should have any illusions that anything in the build is private.
And so we shouldn't generally use .local files for anything which acts
like it is hidden, but actually isn't. They are accurately named .local,
only to be used for machine-specific stuff.

Client IDs are not secret, and so putting them in the .env file is the
right approach.

BTW the .secret files in the build directories should stay there, because
those are build keys which are not included in the app. They are
also in Github as Secrets.